+++ frontend/src/components/drawableMap.js
@@ -6,6 +6,7 @@ import {
   useJsApiLoader,
 } from "@react-google-maps/api";
 import Sidenav from "./sidenav";
+import { sendSinglePolygonToDb, sendToDb, loadPolygon } from "./apiService";
 
 const libraries = ["places", "drawing"];
 
@@ -100,55 +101,6 @@ const DrawableMap = ({ user }) => {
     }
   };
 
-  const sendSinglePolygonToDb = async (coordinates, name, userId) => {
-    try {
-      const response = await fetch(
-        "http://localhost:3000/api/save-single-polygon",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({
-            coordinates,
-            name,
-            userId,
-          }),
-        }
-      );
-
-      if (response.ok) {
-        const result = await response.json();
-        console.log(result.message);
-      } else {
-        const errorData = await response.json();
-        console.error("Failed to save polygon:", errorData.message);
-      }
-    } catch (error) {
-      console.error("Error:", error);
-    }
-  };
-
-  const sendToDb = async () => {
-    try {
-      const response = await fetch("http://localhost:3000/api/fields", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ polygons }),
-      });
-
-      if (response.ok) {
-        console.log("Polygons saved successfully!");
-      } else {
-        console.error("Failed to save polygons");
-      }
-    } catch (error) {
-      console.error("Error:", error);
-    }
-  };
-
   const logPolygons = () => {
     return polygons.map((polygon, index) => ({
       index: index,
@@ -167,30 +119,6 @@ const DrawableMap = ({ user }) => {
       logPolygons();
     }
   }, [polygons, map]);
-
-  const loadFromDB = async (userId) => {
-    try {
-      const response = await fetch(
-        `http://localhost:3000/api/load-polygons/${encodeURIComponent(userId)}`
-      );
-      if (response.ok) {
-        const result = await response.json();
-        const transformedPolygons = result.map((polygon) => ({
-          path: polygon.coordinates.map((coord) => ({
-            lat: coord.lat,
-            lng: coord.lng,
-          })),
-          name: polygon.name,
-        }));
-        setPolygons(transformedPolygons);
-        logPolygons();
-      } else {
-        console.error("Failed to load polygons");
-      }
-    } catch (error) {
-      console.error("Error:", error);
-    }
-  };
 
   const clearMap = () => {
     drawnPolygons.forEach((polygon) => polygon.setMap(null));
@@ -297,6 +225,15 @@ const DrawableMap = ({ user }) => {
       console.error("Error:", error);
     }
   };
+  const loadFromDB = async (userId) => {
+    try {
+      const tPolygons = await loadPolygon(userId);
+      setPolygons(tPolygons);
+      logPolygons();
+    } catch (error) {
+      console.log(error);
+    }
+  };
 
   return isLoaded ? (
     <div style={{ display: "flex" }}>
@@ -305,7 +242,9 @@ const DrawableMap = ({ user }) => {
         user={user}
         logPolygons={logPolygons}
         resetDB={() => resetDB(user.id)}
-        sendToDb={sendToDb}
+        sendToDb={() => {
+          sendToDb(polygons);
+        }}
         loadFromDB={() => loadFromDB(user.id)}
         clearMap={clearMap}
         selectedFieldIndex={selectedFieldIndex}
@@ -317,6 +256,7 @@ const DrawableMap = ({ user }) => {
           center={defaultCenter}
           onLoad={(map) => {
             setMap(map);
+            console.log("i am getting load mutiple times");
           }}
           mapContainerStyle={containerStyle}
         >
